Ya se puede entrar en combate, ganar el combate y perderlo
