+++ src/main/java/dad/gonzacker/pruebasCombate/PruebaApp.java
@@ -1,16 +1,17 @@
 package dad.gonzacker.pruebasCombate;
 
+import dad.gonzacker.controllers.CombateController;
 import javafx.application.Application;
 import javafx.scene.Scene;
 import javafx.stage.Stage;
 
 public class PruebaApp extends Application {
 
-    PruebaController pruebaController = new PruebaController();
+    CombateController combateController = new CombateController();
 
     @Override
     public void start(Stage stage) throws Exception {
-        stage.setScene(new Scene(pruebaController.getRoot()));
+        stage.setScene(new Scene(combateController.getRoot()));
         stage.setTitle("Pruebas");
         stage.setFullScreen(true);
         stage.show();
