--- src/main/resources/fxml/pruebaCombate/pruebaCartas.fxml
@@ -1,87 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import dad.gonzacker.components.UserEntity?>
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.ToolBar?>
-<?import javafx.scene.image.Image?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.layout.BorderPane?>
-<?import javafx.scene.layout.ColumnConstraints?>
-<?import javafx.scene.layout.FlowPane?>
-<?import javafx.scene.layout.GridPane?>
-<?import javafx.scene.layout.HBox?>
-<?import javafx.scene.layout.RowConstraints?>
-<?import org.kordamp.ikonli.javafx.FontIcon?>
-
-<BorderPane fx:id="root" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0" prefWidth="800.0" xmlns="http://javafx.com/javafx/23" xmlns:fx="http://javafx.com/fxml/1">
-   <bottom>
-      <BorderPane BorderPane.alignment="CENTER">
-         <center>
-            <HBox alignment="CENTER_RIGHT" spacing="10.0" BorderPane.alignment="CENTER">
-               <children>
-                  <Label fx:id="energyLabel" text="Label" />
-                  <Label fx:id="maxEnergyLabel" text="Label" />
-                  <Button mnemonicParsing="false" onAction="#onFinishTurnAction" text="Terminar turno" />
-                  <Button mnemonicParsing="false" onAction="#onDrawAction" text="Robar Carta" />
-               </children>
-               <padding>
-                  <Insets bottom="10.0" left="10.0" right="10.0" top="10.0" />
-               </padding>
-            </HBox>
-         </center>
-         <bottom>
-            <FlowPane fx:id="cartasZoneFlowPane" prefHeight="200.0" prefWidth="200.0" BorderPane.alignment="CENTER" />
-         </bottom>
-      </BorderPane>
-   </bottom>
-   <top>
-      <ToolBar prefHeight="40.0" prefWidth="200.0" BorderPane.alignment="CENTER">
-         <items>
-            <Button mnemonicParsing="false">
-               <graphic>
-                  <FontIcon iconLiteral="mdmz-settings" iconSize="20" />
-               </graphic>
-            </Button>
-            <Label fx:id="turnLabel" text="Label" />
-         </items>
-         <padding>
-            <Insets bottom="5.0" left="5.0" right="5.0" top="5.0" />
-         </padding>
-      </ToolBar>
-   </top>
-   <center>
-      <BorderPane styleClass="campo" stylesheets="@../../css/campo.css" BorderPane.alignment="CENTER">
-         <left>
-            <AnchorPane fx:id="battleField" onDragDropped="#onFieldDropped" prefHeight="315.0" prefWidth="346.0" BorderPane.alignment="CENTER">
-               <children>
-                  <UserEntity fx:id="user" escudoActual="0.0" layoutX="73.0" layoutY="60.0">
-                     <imagenEnemigo>
-                        <Image url="@../../images/Programer.png" />
-                     </imagenEnemigo>
-                  </UserEntity>
-               </children>
-            </AnchorPane>
-         </left>
-         <center>
-            <AnchorPane fx:id="userField" onDragDropped="#onFieldDropped" BorderPane.alignment="CENTER" />
-         </center>
-         <right>
-            <GridPane fx:id="enemyGridPane" hgap="10.0" prefHeight="315.0" prefWidth="455.0" vgap="10.0" BorderPane.alignment="CENTER">
-               <columnConstraints>
-                  <ColumnConstraints hgrow="SOMETIMES" maxWidth="1.7976931348623157E308" prefWidth="50.0" />
-                  <ColumnConstraints hgrow="SOMETIMES" maxWidth="1.7976931348623157E308" minWidth="10.0" />
-               </columnConstraints>
-               <rowConstraints>
-                  <RowConstraints minHeight="10.0" vgrow="SOMETIMES" />
-                  <RowConstraints minHeight="10.0" vgrow="SOMETIMES" />
-               </rowConstraints>
-               <BorderPane.margin>
-                  <Insets />
-               </BorderPane.margin>
-            </GridPane>
-         </right>
-      </BorderPane>
-   </center>
-</BorderPane>
