+++ src/main/resources/fxml/recompensaView.fxml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.layout.BorderPane?>
+<?import javafx.scene.layout.VBox?>
+
+<VBox fx:id="root" alignment="CENTER" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" spacing="30.0" stylesheets="@../css/gonzacker.css" xmlns="http://javafx.com/javafx/23" xmlns:fx="http://javafx.com/fxml/1">
+   <children>
+      <Label text="Has recibido:" />
+      <BorderPane fx:id="cartaContainer" />
+      <Label fx:id="moneyLabel" styleClass="money-label" text="Label" />
+      <Button mnemonicParsing="false" onAction="#onContnueAction" text="continuar" />
+   </children>
+</VBox>
